Some recommended and final changes. Javadoc

+++ common/src/main/java/com/epam/esm/common/exception/EntityNotFoundException.java
@@ -4,7 +4,7 @@ import com.epam.esm.common.ErrorDefinition;
 
 public class EntityNotFoundException extends GiftCertificateException {
 
-    public EntityNotFoundException(ErrorDefinition errorDefinition, int id) {
+    public EntityNotFoundException(ErrorDefinition errorDefinition, Long id) {
         super(errorDefinition, id);
     }
 }
